+++ src/collective/collectionfilter/maps/interfaces.py
@@ -1,0 +1,6 @@
+from collective.collectionfilter.interfaces import ICollectionFilterBaseSchema
+
+
+class ICollectionMapsSchema(ICollectionFilterBaseSchema):
+    """Schema for the search filter.
+    """